feat: add distribution for the Tuist App
