+++ .github/workflows/conventional-pr.yml
@@ -1,0 +1,20 @@
+name: conventional-pr
+on:
+  pull_request:
+    paths:
+      - 'app/**'
+    branches:
+      - main
+    types:
+      - opened
+      - edited
+      - synchronize
+
+jobs:
+  lint-pr:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: CondeNast/conventional-pull-request-action@v0.2.0
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
