+++ .mise.toml
@@ -5,10 +5,12 @@ swiftformat = "0.53.3"
 pnpm = "8.15.6"
 node = "20.12.0"
 sourcedocs = "2.0.1"
+create-dmg = "1.2.2"
+"git-cliff" = "2.4.0"
 
 [plugins]
-python = 'https://github.com/asdf-community/asdf-tuist'
 sourcedocs = "https://github.com/tuist/asdf-sourcedocs.git"
+create-dmg = "https://github.com/tuist/asdf-create-dmg.git"
 
 [settings]
 jobs = 6
