+++ .mise/tasks/install
@@ -1,6 +1,8 @@
 #!/bin/bash
-# mise description="Performs additional tasks that are necessary to work in this repository"
+# mise description="Install all necessary dependencies"
 
 set -euo pipefail
 
-pnpm install -C docs/
+pnpm install -C $MISE_PROJECT_ROOT/docs/
+swift package --package-path $MISE_PROJECT_ROOT resolve
+swift package --package-path $MISE_PROJECT_ROOT/app resolve
