+++ Sources/TuistLoader/Loaders/PackageInfoLoader.swift
@@ -1,50 +1,25 @@
 import Foundation
+import Mockable
 import Path
 import TSCUtility
+import TuistSupport
 
 /// Protocol that defines an interface to interact with the Swift Package Manager.
-public protocol SwiftPackageManagerControlling {
-    /// Resolves package dependencies.
-    /// - Parameters:
-    ///   - path: Directory where the `Package.swift` is defined.
-    ///   - printOutput: When true it prints the Swift Package Manager's output.
-    func resolve(at path: AbsolutePath, printOutput: Bool) throws
-
-    /// Updates package dependencies.
-    /// - Parameters:
-    ///   - path: Directory where the `Package.swift` is defined.
-    ///   - printOutput: When true it prints the Swift Package Manager's output.
-    func update(at path: AbsolutePath, printOutput: Bool) throws
-
-    /// Gets the tools version of the package at the given path
+@Mockable
+public protocol PackageInfoLoading {
+    /// Loads the information from the package.
     /// - Parameter path: Directory where the `Package.swift` is defined.
-    /// - Returns: Version of tools.
-    func getToolsVersion(at path: AbsolutePath) throws -> Version
-
-    /// Sets tools version of package to the given value.
-    /// - Parameter path: Directory where the `Package.swift` is defined.
-    /// - Parameter version: Version of tools. When `nil` then the environment’s version will be set.
-    func setToolsVersion(at path: AbsolutePath, to version: Version) throws
-
-    /// Builds a release binary containing release binaries compatible with arm64 and x86.
-    /// - Parameters:
-    ///     - packagePath: Directory where the `Package.swift` is defined.
-    ///     - product: Name of the product to be built.
-    ///     - buildPath: Directory where the intermediary build products should be saved.
-    ///     - outputPath: Directory where the fat binaries should be saved to.
-    func buildFatReleaseBinary(
-        packagePath: AbsolutePath,
-        product: String,
-        buildPath: AbsolutePath,
-        outputPath: AbsolutePath
-    ) throws
+    func loadPackageInfo(at path: AbsolutePath) throws -> PackageInfo
 }
 
-public final class SwiftPackageManagerController: SwiftPackageManagerControlling {
-    let system: Systeming
-    let fileHandler: FileHandling
+public final class PackageInfoLoader: PackageInfoLoading {
+    private let system: Systeming
+    private let fileHandler: FileHandling
 
-    public init(system: Systeming, fileHandler: FileHandling) {
+    public init(
+        system: Systeming = System.shared,
+        fileHandler: FileHandling = FileHandler.shared
+    ) {
         self.system = system
         self.fileHandler = fileHandler
     }
@@ -80,6 +55,17 @@ public final class SwiftPackageManagerController: SwiftPackageManagerControlling
 
         let rawVersion = try system.capture(command).trimmingCharacters(in: .whitespacesAndNewlines)
         return try Version(versionString: rawVersion)
+    }
+
+    public func loadPackageInfo(at path: AbsolutePath) throws -> PackageInfo {
+        let command = buildSwiftPackageCommand(packagePath: path, extraArguments: ["dump-package"])
+
+        let json = try system.capture(command)
+
+        let data = Data(json.utf8)
+        let decoder = JSONDecoder()
+
+        return try decoder.decode(PackageInfo.self, from: data)
     }
 
     public func buildFatReleaseBinary(
