+++ Tests/TuistSupportTests/SwiftPackageManager/SwiftPackageManagerControllerTests.swift
@@ -1,10 +1,10 @@
 import Path
 import TSCUtility
 import TuistCore
-import TuistSupport
+import TuistSupportTesting
 import XcodeGraph
 import XCTest
-@testable import TuistSupportTesting
+@testable import TuistSupport
 
 final class SwiftPackageManagerControllerTests: TuistUnitTestCase {
     private var subject: SwiftPackageManagerController!
@@ -67,90 +67,6 @@ final class SwiftPackageManagerControllerTests: TuistUnitTestCase {
 
         // When / Then
         XCTAssertNoThrow(try subject.setToolsVersion(at: path, to: version!))
-    }
-
-    func test_loadPackageInfo() throws {
-        // Given
-        let path = try temporaryPath()
-        system.succeedCommand(
-            [
-                "swift",
-                "package",
-                "--package-path",
-                path.pathString,
-                "dump-package",
-            ],
-            output: PackageInfo.testJSON
-        )
-
-        // When
-        let packageInfo = try subject.loadPackageInfo(at: path)
-
-        // Then
-        XCTAssertBetterEqual(packageInfo, PackageInfo.test)
-    }
-
-    func test_loadPackageInfo_Xcode14() throws {
-        // Given
-        let path = try temporaryPath()
-        system.succeedCommand(
-            [
-                "swift",
-                "package",
-                "--package-path",
-                path.pathString,
-                "dump-package",
-            ],
-            output: PackageInfo.testJSONXcode14
-        )
-
-        // When
-        let packageInfo = try subject.loadPackageInfo(at: path)
-
-        // Then
-        XCTAssertEqual(packageInfo, PackageInfo.test)
-    }
-
-    func test_loadPackageInfo_alamofire() throws {
-        // Given
-        let path = try temporaryPath()
-        system.succeedCommand(
-            [
-                "swift",
-                "package",
-                "--package-path",
-                path.pathString,
-                "dump-package",
-            ],
-            output: PackageInfo.alamofireJSON
-        )
-
-        // When
-        let packageInfo = try subject.loadPackageInfo(at: path)
-
-        // Then
-        XCTAssertEqual(packageInfo, PackageInfo.alamofire)
-    }
-
-    func test_loadPackageInfo_googleAppMeasurement() throws {
-        // Given
-        let path = try temporaryPath()
-        system.succeedCommand(
-            [
-                "swift",
-                "package",
-                "--package-path",
-                path.pathString,
-                "dump-package",
-            ],
-            output: PackageInfo.googleAppMeasurementJSON
-        )
-
-        // When
-        let packageInfo = try subject.loadPackageInfo(at: path)
-
-        // Then
-        XCTAssertEqual(packageInfo, PackageInfo.googleAppMeasurement)
     }
 
     func test_buildFatReleaseBinary() throws {
