+++ app/.gitignore
@@ -68,3 +68,5 @@ Derived/
 
 ### Tuist managed dependencies ###
 Tuist/.build
+
+build/artifacts
